explora container + auth
