--- Dockerrun.aws.json
@@ -1,12 +1,0 @@
-{
-  "AWSEBDockerrunVersion": "1",
-  "Image": {
-    "Name": "880058582700.dkr.ecr.us-west-1.amazonaws.com/cloud:latest",
-    "Update": "true"
-  },
-  "Ports": [
-    {
-      "ContainerPort": 8999
-    }
-  ]
-}
