+++ dashboard-api/deploy.py
@@ -8,17 +8,21 @@ import boto3
 from botocore.exceptions import ClientError
 
 
-CLOUD_CLUSTER_NAME = "default"
-CLOUD_TASK_DEFINITION = "first-run-task-definition"
-DOMAIN_ID = "/hostedzone/Z3NSW3B4FM6A7X"
+CLUSTER_NAME = "default"
+CLOUD_TASK_DEFINITION = "cloud-task-definition"
+EXPLORA_TASK_DEFINITION = "explora-task-definition"
+CLOUD_CONTAINER_NAME = "cloud-container"
+EXPLORA_CONTAINER_NAME = "explora-container"
 CLOUD_SUBDOMAIN = "{}.cloud.discreetai.com"
 EXPLORA_SUBDOMAIN = "{}.explora.discreetai.com"
+DOMAIN_ID = "/hostedzone/Z3NSW3B4FM6A7X"
+
 
 ecs_client = boto3.client("ecs")
 ec2_client = boto3.resource("ec2")
 route53_client = boto3.client('route53')
 
-def create_new_nodes(repo_id):
+def create_new_nodes(repo_id, api_key):
     """
     Runs new task (node) for ECS cloud cluster. Sets domain of task to be
     `<repo_id>.cloud.discreetai.com by creating an record in Route53.
@@ -31,17 +35,21 @@ def create_new_nodes(repo_id):
         dict: A dictionary holding the public IP address and task ARN of the
             newly created cloud task.
     """
-    cloud_ip_address, cloud_task_arn = _run_new_task(CLOUD_CLUSTER_NAME, \
-        CLOUD_TASK_DEFINITION)
+    cloud_ip_address, cloud_task_arn, explora_ip_address, explora_task_arn = \
+        _run_new_tasks(api_key)
     names = _make_names(repo_id)
-    ip_addresses = [cloud_ip_address]
+    ip_addresses = [cloud_ip_address, explora_ip_address]
     _modify_domains("CREATE", names, ip_addresses)
     return {
         "CloudIpAddress": cloud_ip_address,
-        "CloudTaskArn": cloud_task_arn
+        "CloudTaskArn": cloud_task_arn,
+        "ExploraIpAddress": explora_ip_address,
+        "ExploraTaskArn": explora_task_arn, 
+        "ApiKey": api_key
     }
 
-def stop_nodes(cloud_task_arn, repo_id, cloud_ip_address):
+def stop_nodes(cloud_task_arn, explora_task_arn, repo_id, cloud_ip_address, \
+        explora_ip_address):
     """
     Stop the cloud task with its task ARN and remove the corresponding record 
     in Route53.
@@ -51,12 +59,32 @@ def stop_nodes(cloud_task_arn, repo_id, cloud_ip_address):
         repo_id (str): The repo ID of the repo associated with this task.
         cloud_ip_address (str): The public IP address of the cloud task. 
     """
-    _stop_task(CLOUD_CLUSTER_NAME, cloud_task_arn)
+    _stop_task(cloud_task_arn, explora_task_arn)
     names = _make_names(repo_id)
-    ip_addresses = [cloud_ip_address]
+    ip_addresses = [cloud_ip_address, explora_ip_address]
     _modify_domains("DELETE", names, ip_addresses)
 
-def _create_new_task(cluster_name, task_definition):
+def _run_new_tasks(api_key):
+    """
+    Run new task in the provided cluster with the provided task definition.
+    
+    Args:
+        cluster_name (str): The name of the cluster to run the task in.
+        task_definition (str): The name of the schema of the task to be run.
+    
+    Returns:
+        (str, str): The task ARN and public IP address of the newly created 
+            task.
+    """
+    cloud_task_arn, explora_task_arn = _create_new_task(api_key)
+    cloud_network_interface_id = _get_network_interface_id(cloud_task_arn)
+    explora_network_interface_id = _get_network_interface_id(explora_task_arn)
+    cloud_ip_address = _get_public_ip(cloud_network_interface_id)
+    explora_ip_address = _get_public_ip(explora_network_interface_id)
+    return cloud_ip_address, cloud_task_arn, explora_ip_address, \
+        explora_task_arn
+
+def _create_new_task(api_key):
     """
     Run new task in the provided cluster with the provided task definition.
     
@@ -67,22 +95,58 @@ def _create_new_task(cluster_name, task_definition):
     Returns:
         str: The task ARN of the newly created task.
     """
-    new_task_response = ecs_client.run_task(
-        cluster=cluster_name,
-        taskDefinition=task_definition,
+    cloud_task_response = ecs_client.run_task(
+        cluster=CLUSTER_NAME,
+        taskDefinition=CLOUD_TASK_DEFINITION,
         launchType="FARGATE",
         networkConfiguration = {
             "awsvpcConfiguration": {
                 "subnets": ["subnet-066927cc2aa7d231f"],
                 "assignPublicIp": "ENABLED",
             }
+        },
+        overrides = {
+            "containerOverrides": [{
+                "name": CLOUD_CONTAINER_NAME,
+                "environment": [{
+                    "name": "API_KEY",
+                    "value": api_key
+                }]
+            }]
         }
     )
-    if new_task_response["failures"]:
-        raise Exception(str(new_task_response["failures"]))
-    task_arn = new_task_response["tasks"][0]["taskArn"]
 
-    return task_arn
+    explora_task_response = ecs_client.run_task(
+        cluster=CLUSTER_NAME,
+        taskDefinition=EXPLORA_TASK_DEFINITION,
+        launchType="FARGATE",
+        networkConfiguration = {
+            "awsvpcConfiguration": {
+                "subnets": ["subnet-066927cc2aa7d231f"],
+                "assignPublicIp": "ENABLED",
+            }
+        },
+        overrides = {
+            "containerOverrides": [{
+                "name": EXPLORA_CONTAINER_NAME,
+                "environment": [{
+                    "name": "API_KEY",
+                    "value": api_key
+                }]
+            }]
+        }
+    )
+
+    if cloud_task_response["failures"]:
+        raise Exception(str(new_task_response["failures"]))
+
+    if explora_task_response["failures"]:
+        raise Exception(str(new_task_response["failures"]))
+
+    cloud_task_arn = cloud_task_response["tasks"][0]["taskArn"]
+    explora_task_arn = explora_task_response["tasks"][0]["taskArn"]
+
+    return cloud_task_arn, explora_task_arn
 
 def _get_network_interface_id(task_arn):
     """
@@ -131,24 +195,7 @@ def _get_public_ip(network_interface_id):
     ip_address = ip_address_details["Association"]["PublicIp"]
     return ip_address
 
-def _run_new_task(cluster_name, task_definition):
-    """
-    Run new task in the provided cluster with the provided task definition.
-    
-    Args:
-        cluster_name (str): The name of the cluster to run the task in.
-        task_definition (str): The name of the schema of the task to be run.
-    
-    Returns:
-        (str, str): The task ARN and public IP address of the newly created 
-            task.
-    """
-    task_arn = _create_new_task(cluster_name, task_definition)
-    network_interface_id = _get_network_interface_id(task_arn)
-    ip_address = _get_public_ip(network_interface_id)
-    return ip_address, task_arn
-
-def _stop_task(cluster_name, task_arn):
+def _stop_tasks(cloud_task_arn, explora_task_arn):
     """
     Stop task in the provided cluster with the provided task ARN.
     
@@ -157,8 +204,14 @@ def _stop_task(cluster_name, task_arn):
         task_arn (str): The ARN of the task to be stopped.
     """
     _ = ecs_client.stop_task(
-        cluster=cluster_name,
-        task=task_arn,
+        cluster=CLUSTER_NAME,
+        task=cloud_task_arn,
+        reason="User requested deletion."
+    )
+
+    _ = ecs_client.stop_task(
+        cluster=CLUSTER_NAME,
+        task=explora_task_arn,
         reason="User requested deletion."
     )
 
@@ -217,4 +270,4 @@ def _make_names(repo_id):
     Returns:
         list: The list of domain names for this repo ID.
     """
-    return [CLOUD_SUBDOMAIN.format(repo_id)]
+    return [CLOUD_SUBDOMAIN.format(repo_id), EXPLORA_SUBDOMAIN.format(repo_id)]
