+++ explora/requirements.txt
@@ -1,7 +1,9 @@
 absl-py==0.8.1
 alabaster==0.7.12
+alembic==1.4.2
 appnope==0.1.0
 astor==0.8.0
+async-generator==1.10
 attrs==19.1.0
 autobahn==19.3.3
 Automat==0.7.0
@@ -14,11 +16,14 @@ botocore==1.12.136
 cached-property==1.5.1
 cement==2.8.2
 certifi==2019.3.9
+certipy==0.1.3
+cffi==1.14.0
 chardet==3.0.4
 Click==7.0
 colorama==0.3.9
 constantly==15.1.0
 coremltools==3.3
+cryptography==2.9
 decorator==4.4.0
 defusedxml==0.6.0
 dnspython==1.16.0
@@ -60,10 +65,13 @@ jsonschema==2.6.0
 jupyter==1.0.0
 jupyter-client==5.2.4
 jupyter-console==6.1.0
-jupyter-core==4.4.0
+jupyter-core==4.6.3
+jupyter-telemetry==0.0.5
+jupyterhub==1.1.0
 Keras==2.3.1
 Keras-Applications==1.0.8
 Keras-Preprocessing==1.1.0
+Mako==1.1.2
 Markdown==3.1.1
 MarkupSafe==1.1.1
 mistune==0.8.4
@@ -74,8 +82,10 @@ nbconvert==5.6.1
 nbformat==5.0.4
 notebook==6.0.3
 numpy==1.17.3
+oauthlib==3.1.0
 opt-einsum==3.1.0
 packaging==19.2
+pamela==1.0.0
 pandocfilters==1.4.2
 parsimonious==0.7.0
 parso==0.4.0
@@ -85,19 +95,23 @@ pexpect==4.7.0
 pickleshare==0.7.5
 pluggy==0.13.1
 prometheus-client==0.7.1
-prompt-toolkit==1.0.14
+prompt-toolkit==3.0.5
 protobuf==3.10.0
 ptyprocess==0.6.0
 py==1.8.0
 pyasn1==0.4.5
+pycparser==2.20
 Pygments==2.4.2
 PyHamcrest==1.9.0
 PyInquirer==1.0.3
+pyOpenSSL==19.1.0
 pyparsing==2.4.5
 pyrsistent==0.15.7
 pytest==5.3.1
 python-dateutil==2.8.0
+python-editor==1.0.4
 python-engineio==3.10.0
+python-json-logger==0.1.11
 python-socketio==3.1.2
 pytz==2019.3
 PyYAML==5.1
@@ -106,6 +120,8 @@ qtconsole==4.6.0
 regex==2019.11.1
 requests==2.20.1
 rsa==3.4.2
+ruamel.yaml==0.16.10
+ruamel.yaml.clib==0.2.0
 s3transfer==0.2.0
 scipy==1.2.1
 semantic-version==2.5.0
@@ -122,10 +138,11 @@ sphinxcontrib-htmlhelp==1.0.2
 sphinxcontrib-jsmath==1.0.1
 sphinxcontrib-qthelp==1.0.2
 sphinxcontrib-serializinghtml==1.1.3
+SQLAlchemy==1.3.15
 tb-nightly==1.14.0a20190319
-tensorboard==1.14.0
-tensorflow==1.14.0
-tensorflow-estimator==1.14.0
+tensorboard==1.15.0
+tensorflow==1.15.0
+tensorflow-estimator==1.15.1
 tensorflow-estimator-2.0-preview==1.14.0.dev2019040400
 tensorflow-hub==0.5.0
 tensorflowjs==1.3.1.1
@@ -137,7 +154,7 @@ tornado==6.0.2
 traitlets==4.3.2
 Twisted==19.7.0
 txaio==18.8.1
-urllib3==1.25.7
+urllib3==1.24.3
 uWSGI==2.0.18
 wcwidth==0.1.7
 webencodings==0.5.1
