+++ explora/set_up_jupyter.py
@@ -1,0 +1,7 @@
+import sys
+import os
+
+from notebook.auth.security import set_password
+
+
+set_password(password=sys.argv[1])
